🐛 fix(Entity, MySql): Select attributes and return an Iterator interface
For the Filecahce class, I select the good attributes for it. One more, we have to use an interface iterator to get all files. So, exchanging with a Dev community, I've been told I can use return a PDOStatement type because it uses an interface which is a low-level compared to iterator interface.

Link for Traversable interface: https://www.php.net/manual/fr/class.traversable.php