refactor(): Update for the local storage and create interface
I refactored the updating for the local storage and I created an interface for all Storages Manager.
It's very important to have this interface to force users/developers to use 2 methods very important : getAll(): Entity\Storage and updateId(): void.

I took the opportunity to rename the UserStorage by HomeStorage. This is more logical.